Move packet helpers to own module
